Move environment loading into separate file
This is so it can be required from the cronjob as well as from the
webapp.

On a meta level I'm not entirely happy with how this configuration
loading works right now, it was just cargo culted from a Wikitech
example. I think it might be better in the future to either ignore the
environment and just load a hash from YAML _or_ ignore YAML and just
load the environment in bash so the app doesn't need to worry about it.