Add encoding environment variables
The cronjob on Toolforge is run in a bare environment. When Ruby doesn't
find these environment variables it falls back to running in US-ASCII
mode, which means that any attempts to interact with UTF-8 content gives
the following error:

    invalid byte sequence in US-ASCII

To fix this we need to set these environment variables _before_ running
the cronjob.

Fixes #79